# backend/tests/expected/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "AppSettings": {
        "additionalProperties": true,
        "description": "The full application settings response.",
        "properties": {
          "appVersion": {
            "title": "Appversion",
            "type": "string"
          },
          "dbPath": {
            "title": "Dbpath",
            "type": "string"
          },
          "developerModeEnabled": {
            "title": "Developermodeenabled",
            "type": "boolean"
          },
          "endOfSessionArmourReminderEnabled": {
            "title": "Endofsessionarmourreminderenabled",
            "type": "boolean"
          },
          "gameConnection": {
            "$ref": "#/components/schemas/GameConnection"
          },
          "hotbar": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Hotbar",
            "type": "object"
          },
          "hotbarHooksEnabled": {
            "title": "Hotbarhooksenabled",
            "type": "boolean"
          },
          "lootFilterBlacklist": {
            "items": {
              "type": "string"
            },
            "title": "Lootfilterblacklist",
            "type": "array"
          },
          "mobTrackingMode": {
            "title": "Mobtrackingmode",
            "type": "string"
          },
          "mobTrackingTag": {
            "title": "Mobtrackingtag",
            "type": "string"
          },
          "repairOcrEnabled": {
            "title": "Repairocrenabled",
            "type": "boolean"
          },
          "trifecta": {
            "$ref": "#/components/schemas/TrifectaSettings"
          }
        },
        "required": [
          "gameConnection",
          "hotbarHooksEnabled",
          "repairOcrEnabled",
          "endOfSessionArmourReminderEnabled",
          "developerModeEnabled",
          "mobTrackingMode",
          "mobTrackingTag",
          "hotbar",
          "trifecta",
          "lootFilterBlacklist",
          "dbPath",
          "appVersion"
        ],
        "title": "AppSettings",
        "type": "object"
      },
...
      "GameConnection": {
        "additionalProperties": true,
        "description": "The chat.log connection block of the settings response.",
        "properties": {
          "chatLogPath": {
            "title": "Chatlogpath",
            "type": "string"
          },
          "chatLogValid": {
            "title": "Chatlogvalid",
            "type": "boolean"
          },
          "playerName": {
            "title": "Playername",
            "type": "string"
          }
        },
        "required": [
          "chatLogPath",
          "chatLogValid",
          "playerName"
        ],
        "title": "GameConnection",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OkResponse": {
        "additionalProperties": true,
        "description": "A bare ``{\"ok\": true}`` mutation acknowledgement.",
        "properties": {
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "title": "OkResponse",
        "type": "object"
      },
...
      "OverlayPosition": {
        "additionalProperties": true,
        "description": "The persisted overlay window position (null until first placed).",
        "properties": {
          "x": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "X"
          },
          "y": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Y"
          }
        },
        "title": "OverlayPosition",
        "type": "object"
      },
      "OverlayPositionPatch": {
        "properties": {
          "x": {
            "title": "X",
            "type": "integer"
          },
          "y": {
            "title": "Y",
            "type": "integer"
          }
        },
        "required": [
          "x",
          "y"
        ],
        "title": "OverlayPositionPatch",
        "type": "object"
      },
...
      "SettingsPatch": {
        "properties": {
          "active_trifecta_preset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Trifecta Preset Id"
          },
          "chatlog_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chatlog Path"
          },
          "developer_mode_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Developer Mode Enabled"
          },
          "end_of_session_armour_reminder_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Of Session Armour Reminder Enabled"
          },
          "hotbar": {
            "anyOf": [
              {
                "additionalProperties": {
                  "anyOf": [
                    {
                      "type": "integer"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hotbar"
          },
          "hotbar_hooks_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hotbar Hooks Enabled"
          },
          "loot_filter_blacklist": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Loot Filter Blacklist"
          },
          "mob_tracking_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mob Tracking Mode"
          },
          "mob_tracking_tag": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mob Tracking Tag"
          },
          "player_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Player Name"
          },
          "repair_ocr_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Repair Ocr Enabled"
          },
          "trifecta_presets": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trifecta Presets"
          }
        },
        "title": "SettingsPatch",
        "type": "object"
      },
...
      "TrifectaPreset": {
        "additionalProperties": true,
        "description": "One trifecta loadout preset with its readiness state.",
        "properties": {
          "bigWeaponId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bigweaponid"
          },
          "healId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Healid"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "ready": {
            "title": "Ready",
            "type": "boolean"
          },
          "smallWeaponId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Smallweaponid"
          }
        },
        "required": [
          "id",
          "name",
          "ready"
        ],
        "title": "TrifectaPreset",
        "type": "object"
      },
      "TrifectaSettings": {
        "additionalProperties": true,
        "description": "The trifecta block of the settings response.",
        "properties": {
          "activePresetId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Activepresetid"
          },
          "activePresetName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Activepresetname"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "presets": {
            "items": {
              "$ref": "#/components/schemas/TrifectaPreset"
            },
            "title": "Presets",
            "type": "array"
          },
          "ready": {
            "title": "Ready",
            "type": "boolean"
          }
        },
        "required": [
          "presets",
          "ready"
        ],
        "title": "TrifectaSettings",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "EntropiaOrme API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/settings": {
      "get": {
        "description": "Return full settings including live cache stats.",
        "operationId": "get_settings_api_settings_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppSettings"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Settings",
        "tags": [
          "settings"
        ]
      },
      "patch": {
        "description": "Partial update — only provided fields are written.",
        "operationId": "update_settings_api_settings_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppSettings"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Settings",
        "tags": [
          "settings"
        ]
      }
    },
    "/api/settings/overlay-position": {
      "get": {
        "operationId": "get_overlay_position_api_settings_overlay_position_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OverlayPosition"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Overlay Position",
        "tags": [
          "settings"
        ]
      },
      "put": {
        "operationId": "set_overlay_position_api_settings_overlay_position_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OverlayPositionPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Overlay Position",
        "tags": [
          "settings"
        ]
      }
    },
...
  }
}